requirements.txt
pypiwin32
pygame
Pillow
reportlab
tk